# Conflicts:
#	runners/apex/src/main/java/org/apache/beam/runners/apex/translation/operators/ApexParDoOperator.java
#	runners/core-java/src/main/java/org/apache/beam/runners/core/DoFnRunner.java
#	runners/core-java/src/main/java/org/apache/beam/runners/core/LateDataDroppingDoFnRunner.java
#	runners/core-java/src/main/java/org/apache/beam/runners/core/ProcessFnRunner.java
#	runners/core-java/src/main/java/org/apache/beam/runners/core/PushbackSideInputDoFnRunner.java
#	runners/core-java/src/main/java/org/apache/beam/runners/core/SimpleDoFnRunner.java
#	runners/core-java/src/main/java/org/apache/beam/runners/core/SimplePushbackSideInputDoFnRunner.java
#	runners/core-java/src/main/java/org/apache/beam/runners/core/StatefulDoFnRunner.java
#	runners/core-java/src/test/java/org/apache/beam/runners/core/SimpleDoFnRunnerTest.java
#	runners/core-java/src/test/java/org/apache/beam/runners/core/SimplePushbackSideInputDoFnRunnerTest.java
#	runners/core-java/src/test/java/org/apache/beam/runners/core/StatefulDoFnRunnerTest.java
#	runners/direct-java/src/main/java/org/apache/beam/runners/direct/ParDoEvaluator.java
#	runners/flink/src/main/java/org/apache/beam/runners/flink/metrics/DoFnRunnerWithMetricsUpdate.java
#	runners/flink/src/main/java/org/apache/beam/runners/flink/translation/functions/FlinkStatefulDoFnFunction.java
#	runners/flink/src/main/java/org/apache/beam/runners/flink/translation/wrappers/streaming/DoFnOperator.java
#	runners/flink/src/main/java/org/apache/beam/runners/flink/translation/wrappers/streaming/ExecutableStageDoFnOperator.java
#	runners/flink/src/main/java/org/apache/beam/runners/flink/translation/wrappers/streaming/stableinput/BufferedElements.java
#	runners/flink/src/main/java/org/apache/beam/runners/flink/translation/wrappers/streaming/stableinput/BufferingDoFnRunner.java
#	runners/google-cloud-dataflow-java/worker/src/main/java/org/apache/beam/runners/dataflow/worker/DataflowProcessFnRunner.java
#	runners/google-cloud-dataflow-java/worker/src/main/java/org/apache/beam/runners/dataflow/worker/GroupAlsoByWindowFnRunner.java
#	runners/google-cloud-dataflow-java/worker/src/main/java/org/apache/beam/runners/dataflow/worker/SimpleParDoFn.java
#	runners/google-cloud-dataflow-java/worker/src/main/java/org/apache/beam/runners/dataflow/worker/StreamingKeyedWorkItemSideInputDoFnRunner.java
#	runners/google-cloud-dataflow-java/worker/src/main/java/org/apache/beam/runners/dataflow/worker/StreamingSideInputDoFnRunner.java
#	runners/jet/src/main/java/org/apache/beam/runners/jet/processors/StatefulParDoP.java
#	runners/samza/src/main/java/org/apache/beam/runners/samza/metrics/DoFnRunnerWithMetrics.java
#	runners/samza/src/main/java/org/apache/beam/runners/samza/runtime/DoFnOp.java
#	runners/samza/src/main/java/org/apache/beam/runners/samza/runtime/DoFnRunnerWithKeyedInternals.java
#	runners/samza/src/main/java/org/apache/beam/runners/samza/runtime/SamzaDoFnRunners.java
#	runners/spark/src/main/java/org/apache/beam/runners/spark/structuredstreaming/translation/batch/DoFnRunnerWithMetrics.java
#	runners/spark/src/main/java/org/apache/beam/runners/spark/translation/DoFnRunnerWithMetrics.java
#	runners/spark/src/main/java/org/apache/beam/runners/spark/translation/SparkProcessContext.java

+++ sdks/go/pkg/beam/core/runtime/exec/datasource.go
@@ -262,7 +262,7 @@ func (n *DataSource) Progress() ProgressReportSnapshot {
 // Split takes a sorted set of potential split indices, selects and actuates
 // split on an appropriate split index, and returns the selected split index
 // if successful. Returns an error when unable to split.
-func (n *DataSource) Split(splits []int64, frac float32) (int64, error) {
+func (n *DataSource) Split(splits []int64, frac float64) (int64, error) {
 	if splits == nil {
 		return 0, fmt.Errorf("failed to split: requested splits were empty")
 	}
